requirements.txt
django
django-recaptcha
pillow
six